+++ changelog.md
@@ -1,4 +1,21 @@
 # Changelog
+## [1.1.0] - 2024-06-17
+### Added
+- Users now have a crit rate and crit damage modifier.
+- /profile - you can now view your own or other user profiles and view their stats.
+
+### Changed
+- Changed how damage calculation works.
+- Added many more options to config.json
+
+### Fixed
+- Can no longer go negative when refunding crit vbucks
+- No more variable names in command descriptions
+
+### Planned 
+- Shop - will allow users to purchase items that aid their income and damage in some way.
+
+# Changelog
 ## [1.0.6] - 2024-06-15
 ### Changed
 - More refactoring and config options added